+++ bot/cogs/register.py
@@ -7,8 +7,9 @@ from classes.players import Player, getPlayer
 # Custom modules
 import modules.config as cfg
 from modules.enumerations import PlayerStatus
-from modules.display import channelSend, send, isAlNum
-from modules.exceptions import UnexpectedError, ElementNotFound, CharNotFound, CharInvalidWorld, CharMissingFaction
+from modules.display import channelSend, send
+from modules.tools import isAlNum
+from modules.exceptions import UnexpectedError, ElementNotFound, CharNotFound, CharInvalidWorld, CharMissingFaction, CharAlreadyExists
 from modules.database import update as dbUpdate
 
 
@@ -39,11 +40,12 @@ class registerCog(commands.Cog, name='register'):
         return ctx.channel.id == cfg.discord_ids['register']
 
 
-
-
     @commands.command(aliases=['r'])
     @commands.guild_only()
     async def register(self, ctx, *args):
+        if len(ctx.message.mentions) != 0: # Don't want a mention here
+            await send("REG_INVALID",ctx)
+            return
         try:
             player = getPlayer(ctx.author.id)
         except ElementNotFound:
@@ -52,6 +54,20 @@ class registerCog(commands.Cog, name='register'):
 
         msg = await _register(player, ctx, args)
 
+    @commands.command()
+    @commands.guild_only()
+    async def notify(self, ctx):
+        member = ctx.author
+        notify = member.guild.get_role(cfg.discord_ids["notify_role"])
+        registered = member.guild.get_role(cfg.discord_ids["registered_role"])
+        if notify in member.roles:
+            await member.add_roles(registered)
+            await member.remove_roles(notify)
+            await send("NOTIFY_REMOVED", ctx)
+        else:
+            await member.add_roles(notify)
+            await member.remove_roles(registered)
+            await send("NOTIFY_ADDED", ctx)
 
 def setup(client):
     client.add_cog(registerCog(client))
@@ -84,7 +100,7 @@ async def _register(player, ctx, args):
         return
     wasPlayerRegistered = player.status != PlayerStatus.IS_NOT_REGISTERED # store previous status
     if len(args) == 1 or len(args) == 3: # if 1 or 3 args
-        if len(args) == 1 and args[0].lower()=="help": # =r help displays hel^p
+        if len(args) == 1 and args[0]=="help": # =r help displays hel^p
             await send("REG_HELP",ctx)
             return
         try:
@@ -106,11 +122,14 @@ async def _register(player, ctx, args):
         except CharMissingFaction as e:
             await send("REG_MISSING_FACTION",ctx,e.faction)
             return
+        except CharAlreadyExists as e:
+            await send("REG_ALREADY_EXIST",ctx,e.char, f"<@{e.id}")
+            return
         except UnexpectedError:
             await send("UNKNOWN_ERROR",ctx,"Reg error, check logs")
             return
     if len(args) == 2: # if 2 args, it should be "no account", if not, invalid request. Again, check if update and push db if that's the case
-        if args[0].lower()=="no" and args[1].lower()=="account":
+        if args[0]=="no" and args[1]=="account":
             if not await player.register(None):
                 await send("REG_IS_REGISTERED_NOA",ctx)
                 return